requirements.txt
torch>=1.10.0
torchmetrics==0.11.0
lightning==2.0.0
optuna==3.1.0
tqdm==4.65.0
kaggle==1.5.13
requests~=2.28.2
pandas~=1.5.3
nibabel~=5.0.1
scikit-learn~=1.2.2
tritonclient[grpc]
pydicom~=2.3.1
tifffile~=2023.3.21

numpy~=1.24.2
torchvision~=0.15.1
pydantic~=1.10.6

PyYAML~=6.0

# pip install jupyter